refactor: 공통 함수 훅으로 분리
